+++ rust/pyxel-engine/src/math.rs
@@ -1,4 +1,4 @@
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use std::sync::{LazyLock, Mutex};
 
 use noise::{NoiseFn, Perlin};
@@ -18,15 +18,15 @@ static PERLIN: LazyLock<Mutex<Perlin>> = LazyLock::new(|| {
 });
 
 impl Pyxel {
-    pub fn ceil(x: f64) -> i32 {
-        f64::ceil(x) as i32
+    pub fn ceil(x: f32) -> i32 {
+        f32::ceil(x) as i32
     }
 
-    pub fn floor(x: f64) -> i32 {
-        f64::floor(x) as i32
+    pub fn floor(x: f32) -> i32 {
+        f32::floor(x) as i32
     }
 
-    pub fn sgn(x: f64) -> i32 {
+    pub fn sgn(x: f32) -> i32 {
         if x > 0.0 {
             1
         } else if x < 0.0 {
@@ -36,20 +36,20 @@ impl Pyxel {
         }
     }
 
-    pub fn sqrt(x: f64) -> f64 {
-        f64::sqrt(x)
+    pub fn sqrt(x: f32) -> f32 {
+        f32::sqrt(x)
     }
 
-    pub fn sin(deg: f64) -> f64 {
-        f64::sin(deg * PI / 180.0)
+    pub fn sin(deg: f32) -> f32 {
+        f32::sin(deg * PI / 180.0)
     }
 
-    pub fn cos(deg: f64) -> f64 {
-        f64::cos(deg * PI / 180.0)
+    pub fn cos(deg: f32) -> f32 {
+        f32::cos(deg * PI / 180.0)
     }
 
-    pub fn atan2(y: f64, x: f64) -> f64 {
-        f64::atan2(y, x) * 180.0 / PI
+    pub fn atan2(y: f32, x: f32) -> f32 {
+        f32::atan2(y, x) * 180.0 / PI
     }
 
     pub fn rseed(seed: u32) {
@@ -62,7 +62,7 @@ impl Pyxel {
         RNG.lock().unwrap().random_range(a..=b)
     }
 
-    pub fn rndf(a: f64, b: f64) -> f64 {
+    pub fn rndf(a: f32, b: f32) -> f32 {
         let (a, b) = if a < b { (a, b) } else { (b, a) };
         RNG.lock().unwrap().random_range(a..=b)
     }
@@ -72,7 +72,7 @@ impl Pyxel {
         *PERLIN.lock().unwrap() = perlin;
     }
 
-    pub fn noise(x: f64, y: f64, z: f64) -> f64 {
-        PERLIN.lock().unwrap().get([x, y, z])
+    pub fn noise(x: f32, y: f32, z: f32) -> f32 {
+        PERLIN.lock().unwrap().get([x as f64, y as f64, z as f64]) as f32
     }
 }
